Migrates to react-native 0.29.0
