+++ android/app/src/main/AndroidManifest.xml
@@ -29,6 +29,7 @@
         android:targetSdkVersion="22" />
 
     <application
+      android:name=".MainApplication"
       android:allowBackup="true" android:name="io.eha.goodquestion.GoodQuestionApplication"
       android:label="@string/app_name"
       android:icon="@mipmap/ic_launcher"
