+++ package.json
@@ -14,9 +14,9 @@
     "lodash": "4.11.1",
     "moment": "2.11.2",
     "parse": "1.8.3",
-    "react": "^15.1.0",
-    "react-native": "0.28.0",
     "pubsub-js": "1.5.3",
+    "react": "^15.2.0",
+    "react-native": "^0.29.0",
     "react-native-background-geolocation": "^1.0.0",
     "react-native-background-geolocation-android": "git+ssh://git@github.com/ecohealthalliance/react-native-android.git",
     "react-native-checkbox": "https://github.com/dan-nyanko/react-native-checkbox#master",
